requirements.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0

# API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Model export
onnx>=1.15.0
onnxruntime>=1.16.0

# Training utilities
tqdm>=4.66.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Development
pytest>=7.4.0
httpx>=0.25.0
